backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pillow==10.2.0
mss==9.0.1
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6